website/requirements.txt
wagtail
django-trim